Refactor(user access): Introduce user type that should react to posts and comment but not access the dashboard

+++ controllers/post.controller.js
@@ -23,8 +23,11 @@ export async function getPosts(req, res) {
 export async function addPost(req, res) {
     const { title, intro, body, photo, tags } = req.body
     try {
-        // const photoURL = await upload(req.file?.path)
-        // if (photoURL.secure_url == null) throw "Could not upload image properly.."
+        if (req.user.userType !== 'admin') return res.status(403).json({
+            message: "Admin only are allowed to access this resource",
+            error: true
+        })
+
         let newPost = new Post({
             title,
             intro,
@@ -70,6 +73,11 @@ export async function getOnePost(req, res) {
 
 export async function deletePost(req, res) {
     try {
+        if (req.user.userType !== 'admin') return res.status(403).json({
+            message: "Admin only are allowed to access this resource",
+            error: true
+        })
+
         const { id } = req.params
         const post = await Post.deleteOne({ _id: id })
         if (post) res.status(200).json({ message: "success", data: post })
@@ -83,6 +91,11 @@ export async function deletePost(req, res) {
 
 export async function updateByID(req, res) {
     try {
+        if (req.user.userType !== 'admin') return res.status(403).json({
+            message: "Admin only are allowed to access this resource",
+            error: true
+        })
+
         const { id } = req.params
         const data = req.body
         const post = await Post.findOne({ _id: id })
@@ -90,13 +103,7 @@ export async function updateByID(req, res) {
             res.status(400).json({ message: "There is no post with this ID.." })
             return
         }
-        // let photoURL, picture = {}
-        // if (req.file) {
-        //     photoURL = await upload(req.file.path)
-        //     if (photoURL.secure_url == null) throw "Could not upload image properly.."
-        //     picture = { photo: photoURL.secure_url }
-        // }
-        // return console.log({ data, body: req.body, post, picture })
+
         const updatedPost = Object.assign(post, data)
         const newPost = await Post.findByIdAndUpdate(id, updatedPost, { new: true })
         if (newPost) res.status(200).json({
